% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extracting.R
\name{add_intervals}
\alias{add_intervals}
\title{A function to create a matrix of the metadata extracted from the file
names for adding to a sensor table.}
\usage{
add_intervals(dat, aggregation_unit)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
A function to create a matrix of the metadata extracted from the file
names for adding to a sensor table.
}
\examples{
add(1, 1)
add(10, 1)
}

