% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{add_tables_db}
\alias{add_tables_db}
\title{This function adds tables to a postgreSQL database.}
\usage{
add_tables_db(dbname, port = 5432, user = "postgres")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function adds tables to a postgreSQL database.
}
\examples{
add_tables_db("columbiaBike", port=5432)

}
\seealso{
Other postgresql functions: \code{\link{agg_unit_ok}},
  \code{\link{aggregate_data}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}}, \code{\link{column_types}},
  \code{\link{create_database}}, \code{\link{delete_data}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}},
  \code{\link{kill_pg_connections}},
  \code{\link{list_tables}}, \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
