% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{addZero}
\alias{addZero}
\title{Helper to add zero padding to numbers (mostly for dates/times)}
\usage{
addZero(dat, width = 2)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
Helper to add zero padding to numbers (mostly for dates/times)
}
\examples{
add(1, 1)
add(10, 1)
}

