% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting.R
\name{agg_unit_ok}
\alias{agg_unit_ok}
\title{This function creates a new postgreSQL database .}
\usage{
agg_unit_ok(aggregation_unit)
}
\arguments{
\item{dbname}{Give the database a name.}

\item{port.}{You likely don't need to change this.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
get_sensor_data(tablename="hxi", 
               aggregate = TRUE,
               clean_first = TRUE,
               vars = NULL,
               aggregation_unit="hour",
               grouping_vars = c("subjectid", "sessionid"),
               summarize_vars=c("cadence", "breathing_rate"))


}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}},
  \code{\link{aggregate_data}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}}, \code{\link{column_types}},
  \code{\link{create_database}}, \code{\link{delete_data}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}},
  \code{\link{kill_pg_connections}},
  \code{\link{list_tables}},
  \code{\link{process_hexoskin}},
  \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
