% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting.R
\name{aggregate_data}
\alias{aggregate_data}
\title{This function aggregates data}
\usage{
aggregate_data(dat, aggregation_unit = aggreation_unit,
  summarize_vars = summarize_vars, grouping_vars = grouping_vars)
}
\arguments{
\item{dbname}{Give the database a name.}

\item{aggregation}{unit. examples include "5 min", "2 hour",}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
xxx
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}},
  \code{\link{agg_unit_ok}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}}, \code{\link{column_types}},
  \code{\link{create_database}}, \code{\link{delete_data}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}}, \code{\link{list_tables}},
  \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
