% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregating.R
\name{aggregate_vars}
\alias{aggregate_vars}
\title{This function creates a new postgreSQL database .}
\usage{
aggregate_vars(tablename, aggregation_unit = c("min", "hour", "day"),
  summarize_vars, grouping_vars = c("subjectid", "sessionid"))
}
\arguments{
\item{dbname}{Give the database a name.}

\item{port.}{You likely don't need to change this.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
aggregate_vars(tablename="hxi",
               aggregation_unit="hour",
               grouping_vars = c("subjectid", "sessionid"),
               summarize_vars=c("cadence", "breathing_rate"))
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

