% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{already_uploaded}
\alias{already_uploaded}
\title{Tests if data has already been uploaded}
\usage{
already_uploaded(tablename, filename)
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function tests whether the filename exists in the given table
there is no check to see if the date or data are the same -- based only on
filename
}
\examples{
try(already_uploaded("gps", "BIKE0001_GPS01_S01_150306.gpx"), silent=TRUE)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{create_database}};
  \code{\link{get_connection}};
  \code{\link{upload_postgres}}
}

