% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{backup_database}
\alias{backup_database}
\title{Backup database with pg_dump}
\usage{
backup_database(outpath, outnameprefix, con = ".connection",
  custom_compress = TRUE)
}
\arguments{
\item{outpath}{is the FOLDER you want the backup saved in}

\item{outnameprefix}{is the filename without the file type suffix (suffix is generated
automatically)}

\item{con}{is the name of the database connection}

\item{custom_compress}{is whether or not you want to use a custom-pre-compressed dump
format. If FALSE a standard SQL file is generated.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
backup_database("x:/junk", "mybackup", con=".connection", custom_compress=TRUE)
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}};
  \code{\link{agg_unit_ok}}; \code{\link{aggregate_data}};
  \code{\link{already_uploaded}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}};
  \code{\link{get_sensor_data}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

