% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_gps}
\alias{clean_gps}
\title{A function that cleans HXI data}
\usage{
clean_gps(tablename, dat)
}
\arguments{
\item{tablename}{the table to be cleaned}

\item{data}{to be cleaned}
}
\value{
cleaned data
}
\description{
A function that cleans HXI data
}
\examples{
cleaning_gps("gps", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}},
  \code{\link{clean_data}}, \code{\link{clean_hxi}},
  \code{\link{clean_mae}}, \code{\link{clean_mpm}},
  \code{\link{clean_pdr}}, \code{\link{cleaning_vars}}
}
