% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_mae}
\alias{clean_mae}
\title{A function that cleans HXI data}
\usage{
clean_mae(tablename, dat)
}
\arguments{
\item{tablename}{the table to be cleaned}

\item{data}{to be cleaned}
}
\value{
cleaned data
}
\description{
A function that cleans HXI data
}
\examples{
cleaning_mae("mae", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}};
  \code{\link{clean_data}}; \code{\link{clean_gps}};
  \code{\link{clean_hxi}}; \code{\link{clean_mpm}};
  \code{\link{clean_pdr}}; \code{\link{cleaning_vars}}
}

