% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_mpm}
\alias{clean_mpm}
\title{A function that cleans MPM data}
\usage{
clean_mpm(tablename, dat)
}
\arguments{
\item{tablename}{the table to be cleaned}

\item{data}{to be cleaned}
}
\value{
cleaned data
}
\description{
A function that cleans MPM data
}
\examples{
cleaning_mae("mae", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}};
  \code{\link{clean_data}}; \code{\link{clean_gps}};
  \code{\link{clean_hxi}}; \code{\link{clean_mae}};
  \code{\link{clean_pdr}}; \code{\link{cleaning_vars}}
}

