% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning_vars}
\alias{cleaning_vars}
\title{A function that lists the variables that will be needed in the cleaning step}
\usage{
cleaning_vars(tablename)
}
\arguments{
\item{tablename}{the table to be cleaned}
}
\value{
variables needed for cleaning
}
\description{
A function that lists the variables that will be needed in the cleaning step
}
\examples{
cleaning_vars("hxi")
}
\seealso{
Other xxx: \code{\link{clean_abp}};
  \code{\link{clean_data}}; \code{\link{clean_gps}};
  \code{\link{clean_hxi}}; \code{\link{clean_mae}};
  \code{\link{clean_pdr}}
}

