% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{column_exists}
\alias{column_exists}
\title{Does a column exist in the specified table.}
\usage{
column_exists(tablename, column_names, con = ".connection")
}
\arguments{
\item{tablename}{the postgresql table}

\item{con}{the postgresql connection}
}
\value{
a vector of column names
}
\description{
Does a column exist in the specified table.
}
\examples{
create_database("columbiaBike", port=5432)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{aggregate_vars}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

