% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{createDatabase}
\alias{createDatabase}
\title{This function creates a new postgreSQL database .}
\usage{
createDatabase()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
\code{sum} returns the sum of all the values present in its arguments.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
Other postgresql.functions: \code{\link{addTables}}
}

