% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{delete_data}
\alias{delete_data}
\title{A function to delete data based on a filename}
\usage{
delete_data(tablename, filename, con = ".connection")
}
\value{
user
}
\description{
A function to delete data based on a filename
}
\examples{
#"BIKE0001_GPS01_S01_150306.gpx"
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}},
  \code{\link{agg_unit_ok}}, \code{\link{aggregate_data}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}}, \code{\link{column_types}},
  \code{\link{create_database}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}}, \code{\link{list_tables}},
  \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
