% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{error_report}
\alias{error_report}
\title{xxt}
\usage{
error_report(currentfile_num, currentfile_name, completedfile_names, stage)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
xxt
}
\examples{
add(1, 1)
add(10, 1)
}

