% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filename_ok}
\alias{filename_ok}
\title{A test for whether the input filenames are OK. It is basically a switch
to send to project-specific filename evaluators}
\usage{
filename_ok(filenames, projectid)
}
\arguments{
\item{filenames}{A vector of filenames (not paths)}

\item{projectid}{The projectid grabbed from the shiny app pulldown}
}
\value{
Returns a list with the first item as a vector of TRUE/FALSE about whether
each filename is OK and then a list of the filenames that are NOT Ok.
}
\description{
A test for whether the input filenames are OK. It is basically a switch
to send to project-specific filename evaluators
}
\examples{
prefixes_ok(c("GPS", "AB"))
}
