% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{filename_ok_biking}
\alias{filename_ok_biking}
\title{A test for whether the input filenames are OK for the biking project}
\usage{
filename_ok_biking(filenames)
}
\arguments{
\item{filenames}{A vector of filenames (not paths)}

\item{projectid}{The projectid grabbed from the shiny app pulldown}
}
\value{
Returns a list with the first item as a vector of TRUE/FALSE about whether
each filename is OK and then a list of the filenames that are NOT Ok.
}
\description{
A test for whether the input filenames are OK for the biking project
}
\examples{
prefixes_ok(c("GPS", "AB"))
}
\seealso{
\code{\link{filename_ok}}
}

