% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{find_gaps_assign_session}
\alias{find_gaps_assign_session}
\title{Read GPX files [http://www.topografix.com/gpx.asp] and convert the data to tables;
this function was created by Tomislav Hengl with contributions from Dylan Beaudette
and Pierre Roudier}
\usage{
find_gaps_assign_session(datetime)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
Read GPX files [http://www.topografix.com/gpx.asp] and convert the data to tables;
this function was created by Tomislav Hengl with contributions from Dylan Beaudette
and Pierre Roudier
}
\examples{
add(1, 1)
add(10, 1)
}

