% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{forplot_filter_add_runtime}
\alias{forplot_filter_add_runtime}
\title{Filters out NA based on a variable and adds runtime and date to the 
session ID so that these occur in the plot strip}
\usage{
forplot_filter_add_runtime(dat, thevar)
}
\arguments{
\item{dat}{input data from the database}

\item{thevar}{the variable that will be used for NA filtering}
}
\value{
new dataset with no NA for the var and new sessionid
}
\description{
Filters out NA based on a variable and adds runtime and date to the 
session ID so that these occur in the plot strip
}
\examples{
plotting_hxi("hxi", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}},
  \code{\link{clean_data}}, \code{\link{clean_gps}},
  \code{\link{clean_hxi}}, \code{\link{clean_mae}},
  \code{\link{clean_mpm}}, \code{\link{clean_pdr}},
  \code{\link{cleaning_vars}}, \code{\link{plot_abp}},
  \code{\link{plot_gps}}, \code{\link{plot_hxi}},
  \code{\link{plot_mae}}, \code{\link{plot_mpm}},
  \code{\link{plot_pdr}}, \code{\link{plot_qaqc}},
  \code{\link{sensor_ggplot}}
}
