% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{get_column_names}
\alias{get_column_names}
\title{Get the column names from a PostgreSQL table.}
\usage{
get_column_names(tablename, con = ".connection")
}
\arguments{
\item{tablename}{the postgresql table}

\item{con}{the postgresql connection}
}
\value{
a vector of column names
}
\description{
Get the column names from a PostgreSQL table.
}
\examples{
create_database("columbiaBike", port=5432)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{aggregate_vars}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

