% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{get_connection}
\alias{get_connection}
\title{This function creates the connection to a database}
\usage{
get_connection(dbname, password, host = "localhost", port = 5432,
  user = "postgres")
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function creates the connection to a database
}
\examples{
get_connection(dbname="columbiaBike", host="localhost",
password="spatial", port=5433, user="postgres")
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

