% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{get_row_count}
\alias{get_row_count}
\title{Get a row count from a table}
\usage{
get_row_count(tablename, con = ".connection")
}
\arguments{
\item{tablename}{}

\item{filename}{}
}
\value{
user
}
\description{
Get a row count from a table
}
\examples{
#"BIKE0001_GPS01_S01_150306.gpx"
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{aggregate_vars}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

