% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extracting.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{This function creates a new postgreSQL database .}
\usage{
get_sensor_data(tablename, do_aggregate = FALSE, clean_first = TRUE,
  aggregation_unit = "15 min", xtravars = "all", summarize_vars = NULL,
  grouping_vars = c("datetime", "subjectid", "sessionid"))
}
\arguments{
\item{dbname}{Give the database a name.}

\item{vars.}{list variables for output or "all" for all variables. See details}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
get_sensor_data(tablename="hxi",
               do_aggregate = TRUE,
               clean_first = TRUE,
               xtravars = NULL,
               aggregation_unit="1 hour",
               grouping_vars = c("datetime", "subjectid", "sessionid"),
               summarize_vars=c("cadence", "breathing_rate"))
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}};
  \code{\link{agg_unit_ok}}; \code{\link{aggregate_data}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

