% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R, R/utils.R
\name{getConnection}
\alias{getConnection}
\title{This function creates the connection to a database}
\usage{
getConnection()

getConnection()
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}

\item{sdf}{}
}
\value{
.connection -- which is a global variable

user.
}
\description{
This function creates the connection to a database

xxx
}
\examples{
.connection<<-try(dplyr::src_postgres(dbname="columbiaBike", host="localhost",
password="spatial", port=5433, user="postgres"),silent=TRUE)
add(1, 1)
add(10, 1)
}
\seealso{
Other postgresql.functions: \code{\link{addTables}};
  \code{\link{createDatabase}}
}

