% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_average_to_second}
\alias{hex_average_to_second}
\title{Function for averaging the data from the binary files in the hexoskin to one
second. Specifically for the acceleration X, Y, Z files}
\usage{
hex_average_to_second(.data)
}
\arguments{
\item{.data}{the acceleration data files}
}
\value{
averaged data
}
\description{
Function for averaging the data from the binary files in the hexoskin to one
second. Specifically for the acceleration X, Y, Z files
}
\examples{
xyz
xyz
}
