% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.error}
\alias{is.error}
\title{Test if it's an error}
\usage{
is.error(x)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
Test if it's an error
}
\examples{
add(1, 1)
add(10, 1)
}
\references{
\url{http://adv-r.had.co.nz/Exceptions-Debugging.html}
}

