% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mae}
\alias{plot_mae}
\title{A function that plots HXI data}
\usage{
plot_mae(tablename, dat)
}
\arguments{
\item{tablename}{the table to be ploted}

\item{data}{to be ploted}
}
\value{
ploted data
}
\description{
A function that plots HXI data
}
\examples{
plotting_mae("mae", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}},
  \code{\link{clean_data}}, \code{\link{clean_gps}},
  \code{\link{clean_hxi}}, \code{\link{clean_mae}},
  \code{\link{clean_mpm}}, \code{\link{clean_pdr}},
  \code{\link{cleaning_vars}},
  \code{\link{forplot_filter_add_runtime}},
  \code{\link{plot_abp}}, \code{\link{plot_gps}},
  \code{\link{plot_hxi}}, \code{\link{plot_mpm}},
  \code{\link{plot_pdr}}, \code{\link{plot_qaqc}},
  \code{\link{sensor_ggplot}}
}
