% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{prefixes_ok}
\alias{prefixes_ok}
\title{For now this is deprecated}
\usage{
prefixes_ok(prefixes, allowed = c("GPS", "ABP", "MAE", "MPM", "HXI"))
}
\arguments{
\item{prefixes}{are the prefixes of files}

\item{filenames}{the full list of filenames the user uploaded}

\item{stage}{is the stage of processing ("processing", "uploading",
"pre-screening" right now.)}
}
\value{
user.
}
\description{
For now this is deprecated
}
\examples{
prefixes_ok(c("GPS", "AB"))
}

