% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{process_hexoskin}
\alias{process_hexoskin}
\title{Process hexoskin}
\usage{
process_hexoskin(filepath, filename, fileinfo, metainfilename)
}
\value{
user.
}
\description{
Process hexoskin
}
\examples{
add(1, 1)
add(10, 1)
}
