% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{process_hexoskin_old}
\alias{process_hexoskin_old}
\title{This was the original function for processing the hexoskin and it is replaced
by a function that uses the binary file}
\usage{
process_hexoskin_old(filepath, filename, fileinfo, metainfilename)
}
\value{
user.
}
\description{
This was the original function for processing the hexoskin and it is replaced
by a function that uses the binary file
}
\examples{
add(1, 1)
add(10, 1)
}
