% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{processGPS}
\alias{processGPS}
\title{XXX}
\usage{
processGPS(filepath, filename, fileinfo)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
XXX
}
\examples{
add(1, 1)
add(10, 1)
}

