% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{readGPX}
\alias{readGPX}
\title{Read GPX files [http://www.topografix.com/gpx.asp] and convert the data to tables;
this function was created by Tomislav Hengl with contributions from Dylan Beaudette
and Pierre Roudier}
\usage{
readGPX(gpx.file, metadata = TRUE, bounds = TRUE, waypoints = TRUE,
  tracks = TRUE, routes = TRUE)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
Read GPX files [http://www.topografix.com/gpx.asp] and convert the data to tables;
this function was created by Tomislav Hengl with contributions from Dylan Beaudette
and Pierre Roudier
}
\examples{
add(1, 1)
add(10, 1)
}

