% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{restore_database}
\alias{restore_database}
\title{Restore database with pg_restore}
\usage{
restore_database(dump_path, con = ".connection", create_new = TRUE,
  new_db_name = "newDB")
}
\arguments{
\item{dump_path}{is the full path and file name for the dump file or SQL file}

\item{con}{is the name of the database connection}

\item{create_new}{if this is FALSE then it will create the database using the original DB name. So if the original
was columbiaBike it will create a new columbiaBike. If TRUE then it will create a new database with the name specificied
in new_db_name}

\item{new_db_name}{if creating a new db then this is the name of the new db}
}
\value{
user.
}
\description{
\code{restore_database} will restore the database saved from the backup_database function.
}
\examples{
restore_database("/Users/xyz/backup.dump", con=".connection", create_new = TRUE, new_db_name = "columbiaBike_backup")
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}},
  \code{\link{agg_unit_ok}}, \code{\link{aggregate_data}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}}, \code{\link{column_types}},
  \code{\link{create_database}}, \code{\link{delete_data}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}},
  \code{\link{kill_pg_connections}},
  \code{\link{list_tables}},
  \code{\link{process_hexoskin}},
  \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
