% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShiny}
\alias{runShiny}
\title{This function will run the shiny app.
Thanks to Dean Attali: http://bit.ly/1bDae2R}
\usage{
runShiny(example, package = "sensorDataImport")
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
This function will run the shiny app.
Thanks to Dean Attali: http://bit.ly/1bDae2R
}
\examples{
add(1, 1)
add(10, 1)
}

