% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sensor_ggplot}
\alias{sensor_ggplot}
\title{Create a sensor ggplot with a panel for each session}
\usage{
sensor_ggplot(dat, var, title)
}
\arguments{
\item{dat}{input data from the database}

\item{var}{the variable that will be used for NA filtering}

\item{title}{the plot title}
}
\value{
a ggplot
}
\description{
Create a sensor ggplot with a panel for each session
}
\examples{
plotting_hxi("hxi", dat)
}
\seealso{
Other xxx: \code{\link{clean_abp}},
  \code{\link{clean_data}}, \code{\link{clean_gps}},
  \code{\link{clean_hxi}}, \code{\link{clean_mae}},
  \code{\link{clean_mpm}}, \code{\link{clean_pdr}},
  \code{\link{cleaning_vars}},
  \code{\link{forplot_filter_add_runtime}},
  \code{\link{plot_abp}}, \code{\link{plot_gps}},
  \code{\link{plot_hxi}}, \code{\link{plot_mae}},
  \code{\link{plot_mpm}}, \code{\link{plot_pdr}},
  \code{\link{plot_qaqc}}
}
