% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{testdplyr}
\alias{testdplyr}
\title{test dplyr to make sure that dplyr functions and magrittr pipes are included}
\usage{
testdplyr()
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
test dplyr to make sure that dplyr functions and magrittr pipes are included
}
\examples{
add(1, 1)
add(10, 1)
}

