% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{upload_postgres}
\alias{upload_postgres}
\title{This function is for uploading data to a postgres table}
\usage{
upload_postgres(tablename, data)
}
\arguments{
\item{tablename}{the table name.}

\item{data}{the data to upload}
}
\value{
user
}
\description{
This function is for uploading data to a postgres table
}
\examples{
test
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}};
  \code{\link{agg_unit_ok}}; \code{\link{aggregate_data}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}};
  \code{\link{get_sensor_data}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{valid_connection}}
}

