% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{valid_connection}
\alias{valid_connection}
\title{Test if there is a valid connection.}
\usage{
valid_connection(con = ".connection")
}
\arguments{
\item{dbname}{Give the database a name.}

\item{port.}{You likely don't need to change this.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
create_database("columbiaBike", port=5432)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{aggregate_vars}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}}; \code{\link{list_tables}};
  \code{\link{table_exists}}; \code{\link{upload_postgres}}
}

