% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{add_tables_db}
\alias{add_tables_db}
\title{This function adds tables to a postgreSQL database.}
\usage{
add_tables_db(dbname, port = 5432, user = "postgres")
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function adds tables to a postgreSQL database.
}
\examples{
add_tables_db("columbiaBike", port=5432)
}
\seealso{
Other postgresql.functions: \code{\link{already_uploaded}};
  \code{\link{create_database}};
  \code{\link{get_connection}};
  \code{\link{upload_postgres}}
}

