% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{addTables}
\alias{addTables}
\title{This function adds tables to a postgreSQL database.}
\usage{
addTables()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function adds tables to a postgreSQL database.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
Other postgresql.functions: \code{\link{createDatabase}}
}

