% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{already_uploaded}
\alias{already_uploaded}
\title{Tests if data has already been uploaded}
\usage{
already_uploaded(tablename, filename, con = ".connection")
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function tests whether the filename exists in the given table
there is no check to see if the date or data are the same -- based only on
filename
}
\examples{
try(already_uploaded("gps", "BIKE0001_GPS01_S01_150306.gpx"), silent=TRUE)
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}};
  \code{\link{agg_unit_ok}}; \code{\link{aggregate_data}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}};
  \code{\link{get_sensor_data}}; \code{\link{list_tables}};
  \code{\link{table_exists}};
  \code{\link{upload_postgres}};
  \code{\link{valid_connection}}
}

