% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{body_lines}
\alias{body_lines}
\title{xxt}
\format{\preformatted{ chr [1:65] "\{" "    if (overwrite && append) " ...
}}
\usage{
body_lines
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
xxt
}
\examples{
add(1, 1)
add(10, 1)
}
\keyword{datasets}

