% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseHeader}
\alias{collapseHeader}
\title{Collapses header info into one string separated by pipes (|)
this is done because we don't want each piece of the header
to be its own variable.}
\usage{
collapseHeader(dat, width = NA)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
Collapses header info into one string separated by pipes (|)
this is done because we don't want each piece of the header
to be its own variable.
}
\examples{
add(1, 1)
add(10, 1)
}

