% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{column_types}
\alias{column_types}
\title{What type is the column}
\usage{
column_types(tablename, column_names, con = ".connection")
}
\arguments{
\item{tablename}{the postgresql table}

\item{con}{the postgresql connection}
}
\value{
a vector of column names
}
\description{
What type is the column
}
\examples{
create_database("columbiaBike", port=5432)
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}},
  \code{\link{agg_unit_ok}}, \code{\link{aggregate_data}},
  \code{\link{already_uploaded}},
  \code{\link{backup_database}},
  \code{\link{column_exists}},
  \code{\link{create_database}}, \code{\link{delete_data}},
  \code{\link{get_column_names}},
  \code{\link{get_connection}},
  \code{\link{get_filenames_forSubject}},
  \code{\link{get_filenames}}, \code{\link{get_row_count}},
  \code{\link{get_sensor_data}},
  \code{\link{get_subjectid}}, \code{\link{list_tables}},
  \code{\link{table_exists}},
  \code{\link{upload_postgres}},
  \code{\link{valid_connection}}
}
