% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{create_database}
\alias{create_database}
\title{This function creates a new postgreSQL database .}
\usage{
create_database(dbname, port = 5432, user = "postgres")
}
\arguments{
\item{dbname}{Give the database a name.}

\item{port.}{You likely don't need to change this.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{get_connection}};
  \code{\link{upload_postgres}}
}

