% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{createDatabase}
\alias{createDatabase}
\title{This function creates a new postgreSQL database .}
\usage{
createDatabase(dbname, port = 5432, user = "postgres")
}
\arguments{
\item{dbname}{Give the database a name.}

\item{port.}{You likely don't need to change this.}
}
\value{
user.
}
\description{
\code{createDatabase} will create a new postgresql database.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
Other postgresql.functions: \code{\link{addTables}};
  \code{\link{getConnection}}, \code{\link{getConnection}}
}

