% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{filename_ok}
\alias{filename_ok}
\title{Make sure all the user-selected files have the right prefix.}
\usage{
filename_ok(filenames, projectid)
}
\arguments{
\item{filenames}{the full list of filenames the user uploaded}

\item{prefixes}{are the prefixes of files}

\item{stage}{is the stage of processing ("processing", "uploading",
"pre-screening" right now.)}
}
\value{
user.
}
\description{
Make sure all the user-selected files have the right prefix.
}
\examples{
prefixes_ok(c("GPS", "AB"))
}

