% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_metadata}
\alias{generate_metadata}
\title{A function to create a matrix of the metadata extracted from the file
names for adding to a sensor table.}
\usage{
generate_metadata(fileinfo, n, filename, metainfilename)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
A function to create a matrix of the metadata extracted from the file
names for adding to a sensor table.
}
\examples{
add(1, 1)
add(10, 1)
}

