% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{getConnection}
\alias{getConnection}
\title{This function creates the connection to a database}
\usage{
getConnection(dbname = "columbiaBike", host = "localhost",
  password = "spatial", port = 5432, user = "postgres")
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function creates the connection to a database
}
\examples{
.connection<<-try(dplyr::src_postgres(dbname="columbiaBike", host="localhost",
password="spatial", port=5433, user="postgres"),silent=TRUE)
}
\seealso{
Other postgresql.functions: \code{\link{addTables}};
  \code{\link{createDatabase}}
}

