% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{initiate_processing}
\alias{initiate_processing}
\title{xxt}
\usage{
initiate_processing(filepath, filename, projectid, metainfilename)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
xxt
}
\examples{
add(1, 1)
add(10, 1)
}

