% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{prefixes_ok}
\alias{prefixes_ok}
\title{Make sure all the user-selected files have the right prefix.}
\usage{
prefixes_ok(prefixes, allowed = c("GPS", "ABP", "MAE", "MPM", "HXI"))
}
\arguments{
\item{prefixes}{are the prefixes of files}

\item{filenames}{the full list of filenames the user uploaded}

\item{stage}{is the stage of processing ("processing", "uploading",
"pre-screening" right now.)}
}
\value{
user.
}
\description{
Make sure all the user-selected files have the right prefix.
}
\examples{
prefixes_ok(c("GPS", "AB"))
}

