% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{process_gps}
\alias{process_gps}
\title{xxy}
\usage{
process_gps(filepath, filename, fileinfo)
}
\arguments{
\item{sdf}{}
}
\value{
user.
}
\description{
xxy
}
\examples{
add(1, 1)
add(10, 1)
}

