% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{process_gps}
\alias{process_gps}
\title{xxy}
\usage{
process_gps(filepath, filename, fileinfo)
}
\arguments{
\item{filepath}{is the full filepath with file name}

\item{filename}{is just the filename (this can't be derived from filepathdue to browser restrictions)}

\item{fileinfo}{is a vector of values that are treated as
columns and repeated through the whole file - most likely the split filename but can be anything}
}
\value{
user.
}
\description{
xxy
}
\examples{
add(1, 1)
add(10, 1)
}

