% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_processing.R
\name{process_micropem}
\alias{process_micropem}
\title{xxy}
\usage{
process_micropem(filepath, filename, fileinfo, metainfilename)
}
\value{
user.
}
\description{
xxy
}
\examples{
add(1, 1)
add(10, 1)
}
