% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{upload_postgres}
\alias{upload_postgres}
\title{This function is for uploading data to a postgres table}
\usage{
upload_postgres(tablename, data)
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function is for uploading data to a postgres table
}
\examples{
.connection<<-try(dplyr::src_postgres(dbname="columbiaBike", host="localhost",
password="spatial", port=5433, user="postgres"),silent=TRUE)
}
\seealso{
Other postgresql.functions: \code{\link{add_tables_db}};
  \code{\link{create_database}};
  \code{\link{get_connection}}
}

