% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postgresql.R
\name{valid_connection}
\alias{valid_connection}
\title{This function creates the connection to a database}
\usage{
valid_connection(con = ".connection")
}
\arguments{
\item{dbname}{the database.}

\item{host}{database host, usually 'localhost'}
}
\value{
.connection -- which is a global variable
}
\description{
This function creates the connection to a database
}
\examples{
get_connection(dbname="columbiaBike", host="localhost",
password="spatial", port=5433, user="postgres")
}
\seealso{
Other postgresql functions: \code{\link{add_tables_db}};
  \code{\link{agg_unit_ok}}; \code{\link{aggregate_data}};
  \code{\link{already_uploaded}};
  \code{\link{backup_database}};
  \code{\link{column_exists}}; \code{\link{column_types}};
  \code{\link{create_database}}; \code{\link{delete_data}};
  \code{\link{get_column_names}};
  \code{\link{get_connection}};
  \code{\link{get_row_count}};
  \code{\link{get_sensor_data}}; \code{\link{list_tables}};
  \code{\link{table_exists}}; \code{\link{upload_postgres}}
}

